Phase B7: extract show, request, response handlers into src/commands/
