+++ crates/defra-agent-cli/src/commands/mod.rs
@@ -3,5 +3,8 @@ pub(crate) mod chat;
 pub(crate) mod config;
 pub(crate) mod init;
 pub(crate) mod p2p;
+pub(crate) mod request;
 pub(crate) mod reset;
+pub(crate) mod response;
 pub(crate) mod serve;
+pub(crate) mod show;
