+++ crates/defra-agent-cli/src/main.rs
@@ -214,144 +214,15 @@ async fn main() -> Result<()> {
         Command::Server(args) => commands::serve::serve(args).await,
         Command::Chat(args) => commands::chat::chat(args).await,
         Command::P2p { command } => commands::p2p::dispatch(command).await,
-        Command::Show { command } => match command {
-            ShowCommand::Request(args) => request_show(args).await,
-            ShowCommand::Response(args) => response_show(args).await,
-            ShowCommand::Runtime(args) => show_runtime(args).await,
-        },
+        Command::Show { command } => commands::show::dispatch(command).await,
         Command::Status(args) => status(args).await,
         Command::Diagnose(args) => diagnose(args).await,
         Command::Config { command } => commands::config::dispatch(command).await,
-        Command::Request { command } => match command {
-            RequestCommand::Submit(args) => request_submit(args).await,
-            RequestCommand::Show(args) => request_show(args).await,
-        },
-        Command::Response { command } => match command {
-            ResponseCommand::Show(args) => response_show(args).await,
-            ResponseCommand::Wait(args) => response_wait(args).await,
-        },
+        Command::Request { command } => commands::request::dispatch(command).await,
+        Command::Response { command } => commands::response::dispatch(command).await,
     };
     telemetry.shutdown();
     result
-}
-
-async fn request_submit(args: RequestSubmitArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let agent_did = resolve_agent_did(args.home.as_deref(), args.agent_did.as_deref())?;
-    let content = resolve_request_content(args.content.as_deref(), args.content_file.as_deref())?;
-    let submitted = create_agent_request(
-        &graphql,
-        &agent_did,
-        &content,
-        args.session_id.as_deref(),
-        args.behavior_id.as_deref(),
-        RequestSubmitOptions {
-            temperature: args.temperature,
-            top_p: args.top_p,
-            top_k: args.top_k,
-            max_tokens: args.max_tokens,
-            metadata: args.metadata.clone(),
-        },
-    )
-    .await?;
-    let request_summary = json!({
-        "request_id": submitted.request_id,
-        "session_id": submitted.session_id,
-        "agent_did": submitted.agent_did,
-        "behavior_id": submitted.behavior_id,
-        "temperature": submitted.temperature,
-        "top_p": submitted.top_p,
-        "top_k": submitted.top_k,
-        "max_tokens": submitted.max_tokens,
-        "metadata": submitted.metadata,
-    });
-    if args.no_wait {
-        print_json(&request_summary)?;
-        if let Some(path) = args.output_file.as_deref() {
-            write_json_output_file(path, &request_summary)?;
-        }
-        return Ok(());
-    }
-
-    let response = wait_for_terminal_response(
-        &graphql,
-        &submitted.request_id,
-        args.timeout_secs,
-        args.poll_secs,
-    )
-    .await
-    .with_context(|| format!("waiting for AgentResponse {}", submitted.request_id))?;
-    let mut output = request_summary
-        .as_object()
-        .cloned()
-        .ok_or_else(|| anyhow::anyhow!("request summary was not a JSON object"))?;
-    output.insert("response".to_string(), response);
-    let output = serde_json::Value::Object(output);
-    print_json(&output)?;
-    if let Some(path) = args.output_file.as_deref() {
-        write_json_output_file(path, &output)?;
-    }
-    Ok(())
-}
-
-async fn request_show(args: RequestShowArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let request_id =
-        resolve_request_id(args.request_id.as_deref(), args.request_id_flag.as_deref())?;
-    let query = format!(
-        r#"{{
-            AgentRequest(
-                filter: {{ request_id: {{ _eq: "{request_id}" }} }},
-                order: {{ created_at: DESC }},
-                limit: 1
-            ) {{
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                status
-                lifecycle_state
-                backend_id
-                execution_origin
-                failure_reason
-                retry_count
-                max_retries
-                temperature
-                top_p
-                top_k
-                max_tokens
-                metadata
-                created_at
-                claimed_at
-                deadline
-            }}
-        }}"#,
-        request_id = escape_graphql_string(&request_id),
-    );
-    let response = post_graphql(&graphql, &query).await?;
-    print_json(&response)?;
-    Ok(())
-}
-
-async fn response_show(args: ResponseShowArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let request_id =
-        resolve_request_id(args.request_id.as_deref(), args.request_id_flag.as_deref())?;
-    let query = response_query(&request_id);
-    let response = post_graphql(&graphql, &query).await?;
-    print_json(&response)?;
-    Ok(())
-}
-
-async fn response_wait(args: ResponseWaitArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let request_id =
-        resolve_request_id(args.request_id.as_deref(), args.request_id_flag.as_deref())?;
-    let response =
-        wait_for_terminal_response(&graphql, &request_id, args.timeout_secs, args.poll_secs)
-            .await?;
-    print_json(&response)?;
-    Ok(())
 }
 
 async fn status(args: StatusArgs) -> Result<()> {
@@ -362,14 +233,6 @@ async fn status(args: StatusArgs) -> Result<()> {
     Ok(())
 }
 
-
-async fn show_runtime(args: RuntimeShowArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let agent_did = resolve_agent_did(args.home.as_deref(), args.agent_did.as_deref())?;
-    let output = load_runtime_status_output(args.home.as_deref(), &graphql, &agent_did).await?;
-    print_json(&output)?;
-    Ok(())
-}
 
 async fn diagnose(args: DiagnoseArgs) -> Result<()> {
     let home_dir = resolve_home_dir(args.home.as_deref());
@@ -547,7 +410,7 @@ async fn diagnose(args: DiagnoseArgs) -> Result<()> {
     Ok(())
 }
 
-async fn load_runtime_status_output(
+pub(crate) async fn load_runtime_status_output(
     home: Option<&Path>,
     graphql: &str,
     agent_did: &str,
@@ -2434,7 +2297,7 @@ pub(crate) fn resolve_agent_did(home: Option<&Path>, explicit: Option<&str>) -> 
     Ok(format!("did:defra-agent:{DEFAULT_AGENT_NAME}"))
 }
 
-fn resolve_request_id(positional: Option<&str>, flag: Option<&str>) -> Result<String> {
+pub(crate) fn resolve_request_id(positional: Option<&str>, flag: Option<&str>) -> Result<String> {
     let positional = positional.map(str::trim).filter(|value| !value.is_empty());
     let flag = flag.map(str::trim).filter(|value| !value.is_empty());
     match (positional, flag) {
@@ -2517,7 +2380,7 @@ pub(crate) fn resolve_task_prompt(prompt: Option<&str>, prompt_file: Option<&Pat
     }
 }
 
-fn resolve_request_content(content: Option<&str>, content_file: Option<&Path>) -> Result<String> {
+pub(crate) fn resolve_request_content(content: Option<&str>, content_file: Option<&Path>) -> Result<String> {
     match (content, content_file) {
         (Some(_), Some(path)) => anyhow::bail!(
             "provide either --content or --content-file, not both ({})",
